# Embeddings: use SIMD instructions for search on arm64