--- a/enterprise/internal/embeddings/dot_arm64.s
+++ b/enterprise/internal/embeddings/dot_arm64.s
@@ -1,6 +1,5 @@
 #include "textflag.h"
 
-#define BLOCKSIZE $16
 
 TEXT ·dotSIMD(SB), NOSPLIT, $0-56
 	// Offsets based on slice header offsets.
@@ -9,30 +8,35 @@ TEXT ·dotSIMD(SB), NOSPLIT, $0-56
 	MOVD b_base+24(FP), R5
 	MOVD a_len+8(FP), R6
 
+	ADD R4, R6, R6 // end pointer
 
 	// Zero V0, which will store 4 packed 32-bit sums
 	VEOR V0.B16, V0.B16, V0.B16
 
+#define BLOCKSIZE $16
+
 blockloop:
-    CMP BLOCKSIZE, R6
-    BLT reduce
+	CMP R4, R6
+	BEQ reduce
 
-    VLD1 (R4), [V1.B16]
-    VLD1 (R5), [V2.B16]
+	// Load 16 bytes from each slice
+	VLD1.P 16(R4), [V1.B16]
+	VLD1.P 16(R5), [V2.B16]
 
-    // The following instruction is not supported
-    // by the go assembler, so use the binary format.
+	// The following instruction is not supported by the go assembler, so use
+	// the binary format. It would be the equivalent of the following instruction:
+	//
     // VSDOT V1.B16, V2.B16, V0.S4
-    WORD $0x4E829420
+	//
+	// I generated the binary form of the instruction using this godbolt setup:
+	// https://godbolt.org/z/3jPohn4dn
+	WORD $0x4E829420
 
-	ADD BLOCKSIZE, R4
-	ADD BLOCKSIZE, R5
-	SUB BLOCKSIZE, R6
 	JMP blockloop
 
 reduce:
-    VADDV V0.S4, V0
-    VMOV V0.S[0], R6
-    MOVD R6, ret+48(FP)
+	VADDV V0.S4, V0
+	VMOV V0.S[0], R6
+	MOVD R6, ret+48(FP)
 	RET