--- a/enterprise/internal/embeddings/dot_portable.go
+++ b/enterprise/internal/embeddings/dot_portable.go
@@ -1,4 +1,4 @@
-//go:build !amd64
+//go:build !amd64 && !arm64
 
 package embeddings