This adds an SIMD version of embeddings search for the arm64 arch. It uses the builtin dot product vector instruction (SDOT), which works beautifully for our use case.

Not high priority since this would only actually be useful for searching embeddings in App and/or local dev, but I was playing around with ARM assembly this weekend and this was the motivating problem.

## Test plan

Added a bunch of tests comparing it to the naive version. Quicktests exercise many cases. Behind the same environment variable as SIMD instructions on amd64, so this is low risk. Benchmark below.

```
name                               old time/op              new time/op              delta
SimilaritySearch/numWorkers=1-10                433ms ± 1%                93ms ± 0%   -78.41%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=2-10                224ms ± 1%                47ms ± 0%   -78.80%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=4-10                115ms ± 0%                24ms ± 0%   -79.23%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=8-10               58.8ms ± 3%              13.0ms ± 1%   -77.81%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=16-10              66.2ms ± 2%              13.6ms ± 1%   -79.45%  (p=0.000 n=10+10)

name                               old embeddings/s         new embeddings/s         delta
SimilaritySearch/numWorkers=1-10                2.31M ± 1%              10.71M ± 0%  +363.19%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=2-10                4.47M ± 1%              21.11M ± 0%  +371.71%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=4-10                8.73M ± 0%              42.01M ± 0%  +381.47%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=8-10                17.0M ± 3%               76.6M ± 1%  +350.52%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=16-10               15.1M ± 2%               73.5M ± 1%  +386.52%  (p=0.000 n=10+10)

name                               old embeddings/s/worker  new embeddings/s/worker  delta
SimilaritySearch/numWorkers=1-10                2.31M ± 1%              10.71M ± 0%  +363.19%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=2-10                2.24M ± 1%              10.55M ± 0%  +371.71%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=4-10                2.18M ± 0%              10.50M ± 0%  +381.47%  (p=0.000 n=10+8)
SimilaritySearch/numWorkers=8-10                2.13M ± 3%               9.58M ± 1%  +350.52%  (p=0.000 n=10+9)
SimilaritySearch/numWorkers=16-10                944k ± 2%               4593k ± 1%  +386.52%  (p=0.000 n=10+10)
```
<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles -->