@eseliger 2023-05-08
>  This may be a silly question, does our CI test on arm64 machines?

Not silly, and we don't, it's all amd64